v1/requirements.txt
# App / LLM
streamlit>=1.30.0
openai>=1.0.0

# PDFs
PyPDF2>=3.0.0
reportlab>=4.0.0

# Word docs
python-docx>=1.1.0

# Spreadsheets
openpyxl>=3.1.0

# Slides (JS-based; see note below)
# pptxgenjs is a Node.js package, not pip-installable.